Bruce Sherwood Updates
visual.py (line 5): Comment out "from array import array" which is not
used
but conflicts with "array" in the case of "from numpy import *".
visual.py (line 348): Made mouse.pos and mouse.pickpos be vectors, not
tuples.
mouse.pos has to be a vector because existing programs expect to be able
to do vector calculations with it.
visual.py (line 556): Fixed a sign error; rotations went clockwise
instead of counterclockwise,
and this rotate function must return the new vector, not change self.
visual.py (line 789): object.rotate must be done in server, because the
user program reads axis and up. Fixed.

In addition, I made minor changes throughout visual.py so that it works
with either Python 2 or Python 3.